+++ source/Draft/Endpoints/EndpointVerificationStrategy.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.Serialization;
 using System.Threading.Tasks;
 
 using Draft.Exceptions;
@@ -10,9 +11,11 @@ namespace Draft.Endpoints
     /// <summary>
     ///     Represents a strategy for verifying etcd enpoint availability.
     /// </summary>
+    [Serializable, DataContract]
     public abstract class EndpointVerificationStrategy
     {
 
+        [IgnoreDataMember]
         internal static EndpointVerificationStrategy Default
         {
             get { return None; }
@@ -27,15 +30,22 @@ namespace Draft.Endpoints
 
         #region Built-in strategies
 
+        [IgnoreDataMember]
         private static readonly Lazy<EndpointVerificationStrategy> LazyAll = new Lazy<EndpointVerificationStrategy>(() => new VerificationStrategyAll());
 
+        [IgnoreDataMember]
         private static readonly Lazy<EndpointVerificationStrategy> LazyAny = new Lazy<EndpointVerificationStrategy>(() => new VerificationStrategyAny());
 
+        [IgnoreDataMember]
+        private static readonly Lazy<EndpointVerificationStrategy> LazyClusterMembers = new Lazy<EndpointVerificationStrategy>(() => new VerificationStrategyClusterMembers());
+
+        [IgnoreDataMember]
         private static readonly Lazy<EndpointVerificationStrategy> LazyNone = new Lazy<EndpointVerificationStrategy>(() => new VerificationStrategyNone());
 
         /// <summary>
         ///     Verify all supplied endpoints are responding.
         /// </summary>
+        [IgnoreDataMember]
         public static EndpointVerificationStrategy All
         {
             get { return LazyAll.Value; }
@@ -44,14 +54,25 @@ namespace Draft.Endpoints
         /// <summary>
         ///     Verify any supplied endpoints are responding.
         /// </summary>
+        [IgnoreDataMember]
         public static EndpointVerificationStrategy Any
         {
             get { return LazyAny.Value; }
         }
 
         /// <summary>
+        ///     Similar verification as <see cref="Any" /> but also adds verified cluster members to the <see cref="EndpointPool" />.
+        /// </summary>
+        [IgnoreDataMember]
+        public static EndpointVerificationStrategy ClusterMembers
+        {
+            get { return LazyClusterMembers.Value; }
+        }
+
+        /// <summary>
         ///     Doesn't do any endpoint verification.
         /// </summary>
+        [IgnoreDataMember]
         public static EndpointVerificationStrategy None
         {
             get { return LazyNone.Value; }