+++ source/Draft/Endpoints/RoutingStrategy.Random.cs
@@ -1,18 +1,27 @@
 ﻿using System;
 using System.Linq;
+using System.Runtime.Serialization;
 
 namespace Draft.Endpoints
 {
+    [Serializable, DataContract]
     internal sealed class RoutingStrategyRandom : EndpointRoutingStrategy
     {
 
-        private readonly IRandom _random;
+        [NonSerialized, IgnoreDataMember]
+        private IRandom _random;
 
         public RoutingStrategyRandom() : this(StaticRandom.Instance) {}
 
         public RoutingStrategyRandom(IRandom random)
         {
             _random = random ?? StaticRandom.Instance;
+        }
+
+        [OnDeserialized]
+        private void OnDeserialized(StreamingContext ignored)
+        {
+            _random = _random ?? StaticRandom.Instance;
         }
 
         public override Endpoint Select(string key, Endpoint[] endpoints)