+++ source/Draft/Endpoints/RoutingStrategy.RoundRobin.cs
@@ -1,12 +1,15 @@
 ﻿using System;
 using System.Linq;
+using System.Runtime.Serialization;
 using System.Threading;
 
 namespace Draft.Endpoints
 {
+    [Serializable, DataContract]
     internal sealed class RoutingStrategyRoundRobin : EndpointRoutingStrategy
     {
 
+        [DataMember(Order = 1)]
         private int _next = -1;
 
         public override Endpoint Select(string key, Endpoint[] endpoints)