+++ source/Draft/Endpoints/VerificationStrategy.None.cs
@@ -1,10 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.Serialization;
 using System.Threading.Tasks;
 
 namespace Draft.Endpoints
 {
+    [Serializable, DataContract]
     internal sealed class VerificationStrategyNone : EndpointVerificationStrategy
     {