+++ source/Draft/Requests/Cluster/GetMembersRequest.cs
@@ -14,7 +14,7 @@ namespace Draft.Requests.Cluster
     internal class GetMembersRequest : BaseRequest, IGetMembersRequest
     {
 
-        public GetMembersRequest(IEtcdClient etcdClient, EndpointPool endpointPool, params string[] pathParts) 
+        public GetMembersRequest(IEtcdClient etcdClient, EndpointPool endpointPool, params string[] pathParts)
             : base(etcdClient, endpointPool, pathParts) {}
 
         public async Task<IClusterMember[]> Execute()