+++ source/Draft/Responses/Statistics/IFollowerCounts.cs
@@ -1,27 +1,23 @@
 ﻿using System;
 using System.Linq;
-using System.Runtime.Serialization;
 
 namespace Draft.Responses.Statistics
 {
     /// <summary>
     ///     Follower send counts
     /// </summary>
-    [DataContract]
-    internal class FollowerCounts : IFollowerCounts
+    public interface IFollowerCounts
     {
 
         /// <summary>
         ///     Count of unsuccessful sends
         /// </summary>
-        [DataMember(Name = "fail")]
-        public long Fail { get; set; }
+        long Fail { get; }
 
         /// <summary>
         ///     Count of successful sends
         /// </summary>
-        [DataMember(Name = "success")]
-        public long Success { get; set; }
+        long Success { get; }
 
     }
 }