+++ source/Draft/Responses/Statistics/ILeaderInfo.cs
@@ -1,34 +1,28 @@
 ﻿using System;
 using System.Linq;
-using System.Runtime.Serialization;
 
 namespace Draft.Responses.Statistics
 {
     /// <summary>
     ///     Statistical information for a member server's leader
     /// </summary>
-    [DataContract]
-    internal class LeaderInfo : ILeaderInfo
+    public interface ILeaderInfo
     {
 
         /// <summary>
         ///     Id of the current leader
         /// </summary>
-        [DataMember(Name = "leader")]
-        public string Leader { get; set; }
+        string Leader { get; }
 
         /// <summary>
         ///     Time when the leader was started (if available)
         /// </summary>
-        [DataMember(Name = "startTime")]
-        public DateTime? StartTime { get; set; }
+        DateTime? StartTime { get; }
 
         /// <summary>
         ///     String representation of the amount of time the leader has been leader
         /// </summary>
-        [DataMember(Name = "uptime")]
-        public string Uptime { get; set; }
+        string Uptime { get; }
 
     }
-
 }